requirements.txt
appdirs==1.4.0
future==0.16.0
nltk==3.2.2
oauthlib==2.0.1
packaging==16.8
pyparsing==2.1.10
python-twitter==3.2.1
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
